90 hak akses data peusahaan

+++ app/Http/Controllers/Master/General/CompanyController.php
@@ -17,12 +17,14 @@ class CompanyController extends Controller
 
     public function index(): View
     {
+        $this->authorize('view', Company::class);
         return view('pages.general-master-data.company.index');
     }
 
 
     public function data(): JsonResponse
     {
+        $this->authorize('view', Company::class);
         $company = Company::paginate(self::$perPage);
         return response()->json($company);
     }
@@ -30,12 +32,13 @@ class CompanyController extends Controller
 
     public function search(Request $request): JsonResponse
     {
+        $this->authorize('view', Company::class);
         $search = $request->input('search');
         $company = Company::paginate(self::$perPage);
 
 
         if (!empty($search)) {
-            $company->where('name', 'like', '%'.$search.'%');
+            $company->where('name', 'like', '%' . $search . '%');
         }
 
         $data = $company->paginate(self::$perPage);
@@ -45,18 +48,21 @@ class CompanyController extends Controller
 
     public function store(CompanyRequest $request): JsonResponse
     {
+        $this->authorize('create', Company::class);
         return response()->json(Company::create($request->validated()));
     }
 
 
     public function edit(Company $company): JsonResponse
     {
+        $this->authorize('edit', $company);
         return response()->json($company);
     }
 
 
     public function update(CompanyRequest $request, Company $company): JsonResponse
     {
+        $this->authorize('edit', $company);
         return response()->json($company->update($request->validated()));
     }
 
@@ -66,6 +72,7 @@ class CompanyController extends Controller
      */
     public function destroy(Request $request, Company $company): JsonResponse
     {
+        $this->authorize('delete', $company);
         $implodeID = implode(',', $request->get('id'));
         $explodeID = explode(',', $implodeID);
         $company->whereIn('id', $explodeID)->delete();
